pytest-embedded-wokwi/pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pytest-embedded-wokwi"
authors = [
    {name = "Fu Hanxi", email = "fuhanxi@espressif.com"},
    {name = "Uri Shaked", email = "uri@wokwi.com"},
    {name = "Jakub Andrysek", email = "jakub.andrysek@espressif.com"},
]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python",
    "Topic :: Software Development :: Testing",
]
dynamic = ["version", "description"]
requires-python = ">=3.10"

dependencies = [
    "pytest-embedded~=1.17.0a2",
    "wokwi-client>=0.2.0",
]

[project.optional-dependencies]
idf = [
    "pytest-embedded-idf~=1.17.0a2",
]

[project.urls]
homepage = "https://github.com/espressif/pytest-embedded"
repository = "https://github.com/espressif/pytest-embedded"
documentation = "https://docs.espressif.com/projects/pytest-embedded/en/latest/"
changelog = "https://github.com/espressif/pytest-embedded/blob/main/CHANGELOG.md"
